pyproject.toml
[project]
name = "ai-trader-expansion"
version = "0.1.0"
description = "AI Multi-Trader Expansion - 7戦略・6ブローカー対応株式トレーディング自動化"
requires-python = ">=3.12"
dependencies = [
    # データ取得
    "yfinance>=0.2.36",
    "requests>=2.31.0",
    # データ処理
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    # 機械学習
    "lightgbm>=4.3.0",
    "scikit-learn>=1.4.0",
    # テクニカル指標
    "pandas-ta>=0.3.14b",
    # Web APIサーバー
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
]
jquants = [
    "jquants-api-client>=2.0.0",
]
ml = [
    "darts>=0.28.0",
    "statsforecast>=1.7.0",
    "neuralforecast>=1.7.0",
    "skforecast>=0.12.0",
]
finrl = [
    "finrl>=0.3.6",
    "gymnasium>=0.29.0",
]
research = [
    "qlib>=0.9.0",
]
all = [
    "ai-trader-expansion[jquants,ml,finrl,research]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "TCH"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
